change: reformatted stashes
stashes now show time ago stats instead of which stash it was. so instead of stash{0} its now 34m On main.

also gave it some formatting so that the name/title of the stash and the branch is inline. also removed the coloring from the name/title in favour of readability